add: Extension function for livedata observation
added extension function so that we request data from API and using observer show that in recycler view. Earlier it was taking data from API and then storing it in a local data and then was showing it. This functionality removes that step and now data is shown after being called from API when database is empty